Long Type 을 Int Type으로 변경

+++ Back/mesbiens/src/main/java/mesbiens/community/quiz/controller/CreateQuizRequest.java
@@ -12,7 +12,7 @@ import mesbiens.community.quiz.vo.QuizVo;
 @Setter
 public class CreateQuizRequest {
 	
-	private Long memberId;
+	private int memberId;
     private QuizVo quizVo;
     private String quizQuestion;
     private String quizCorrectAnswer;
@@ -27,11 +27,11 @@ public class CreateQuizRequest {
     // 기본 생성자
     public CreateQuizRequest() {}
     
-    public Long getMemberId() {
+    public int getMemberId() {
         return memberId;
     }
 
-    public void setMemberId(Long memberId) {
+    public void setMemberId(int memberId) {
         this.memberId = memberId;
     }
