+++ Back/mesbiens/src/main/java/mesbiens/community/quiz/controller/QuizController.java
@@ -34,12 +34,12 @@ public class QuizController {
 
     // 특정 퀴즈 조회
     @GetMapping("/{quizId}")
-    public Optional<QuizVo> getQuizById(@PathVariable Long quizId) {
+    public Optional<QuizVo> getQuizById(@PathVariable int quizId) {
         return quizService.getQuizById(quizId);
     }
 
     @PostMapping("/create")
-    public ResponseEntity<QuizVo> createQuiz(@RequestParam Long memberId, @RequestBody QuizVo quizVo) {
+    public ResponseEntity<QuizVo> createQuiz(@RequestParam int memberId, @RequestBody QuizVo quizVo) {
         
     	 QuizVo createdQuiz = quizService.createQuiz(memberId, quizVo);
     	// 클라이언트에서 전달한 CreateQuizRequest 객체를 처리
@@ -52,7 +52,7 @@ public class QuizController {
 
     // 퀴즈 수정
     @PutMapping("/update/{quizId}")
-    public ResponseEntity<QuizVo> updateQuiz(@PathVariable Long quizId, @RequestBody QuizVo updatedQuizVo) {
+    public ResponseEntity<QuizVo> updateQuiz(@PathVariable int quizId, @RequestBody QuizVo updatedQuizVo) {
         try {
             QuizVo updatedQuiz = quizService.updateQuiz(quizId, updatedQuizVo);
             return ResponseEntity.ok(updatedQuiz);
@@ -63,7 +63,7 @@ public class QuizController {
 
     // 퀴즈 삭제
     @DeleteMapping("/delete/{quizId}")
-    public ResponseEntity<String> deleteQuiz(@PathVariable Long quizId) {
+    public ResponseEntity<String> deleteQuiz(@PathVariable int quizId) {
         boolean deleted = quizService.deleteQuiz(quizId);
         if (deleted) {
             return ResponseEntity.ok("Quiz deleted successfully");
